Artists features added, Artists can now be connected with Genres, more indepth details, start with Albums made.
minor fixes have been added to the Artist CRUD features,
You are now able to add Genres to artists, and to see all the genres of
an artist.
You can add a genre to an artist through a dropdownlist, this has a
searchbar, and only displays genres which the artist doesn't make yet.
Controller for Album has been made, and Data Access Object for Album has
been made.